Create code_1
This code uses the Node.js https module to make a GET request to the Yahoo Finance API, passing in the VISA stock symbol as a parameter. The response is received in chunks and appended to the data variable. Once the entire response is received, it uses a regular expression to extract the current stock price from the response and logs it to the console.

Please note: This is a code snippet and to run the whole code you need to install https package in node js.